\name{rbioc2011-package}
\alias{rbioc2011-package}
\alias{rbioc2011}
\docType{package}
\title{
  Example package for showing gitsvn at BioC 2011 (Manchester)
}
\description{
  TODO
}
\details{
  \tabular{ll}{
    Package: \tab rbioc2011\cr
    Type: \tab Package\cr
    Version: \tab 0.1\cr
    Date: \tab 2011-11-29\cr
    License: \tab GPL-2\cr
  }
}
\author{
  Tim Yates\cr
  Maintainer: Tim Yates <tyates(at)picr.man.ac.uk>
}
\keyword{ package }
